move function to utils
